requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
pydantic-settings
alembic
python-multipart
sqlitecloud
sqlalchemy-sqlitecloud
numpy
structlog
python-jose[cryptography]
bleach
passlib
